Update problemOne.java
Forgot to change return type from long to int

+++ problemOne.java
@@ -4,7 +4,7 @@ public class problemOne {
     }
 
 
-    public static long findSum() {
+    public static int findSum() {
         int result = 0;
         for (int i = 0; i < 1000; i++) {
             if(checkMultiple(i)) {
